Save Now OK Again
By deleting components from the Tent Object, I deleted a component that other elements tracked. So that caused the isse.

By reconnecting them with an existing component of the Tent the bug was fixed
