--- Project Smash/Assets/_Scripts/Checkpoint/PlayerPositionCheckpoint.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 708a4261291fd8f439b6eacdd5896cc2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
